fix: clarify JSM connector scoping
